% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasetExample}
\alias{datasetExample}
\title{Alterations and expression data from 50 lung adenocarcinoma samples obtained from TCGA studies.}
\format{
List with two matrices providing alterations and expression data for a set of selected genes.
}
\source{
Cancer Genome Atlas Research Network. "Comprehensive molecular profiling of lung adenocarcinoma." Nature 511.7511 (2014): 543.
}
\usage{
data(datasetExample)
}
\value{
List with two matrices providing alterations and expression data from 50 lung adenocarcinoma samples for a set of selected genes.
}
\description{
Alterations and expression data for selected genes from 50 lung adenocarcinoma samples from Cancer Genome Atlas Research Network. "Comprehensive molecular profiling of lung adenocarcinoma." Nature 511, no. 7511 (2014): 543.
}
